Axios functions

+++ app/api/issues/[id]/route.ts
@@ -4,12 +4,14 @@ import prisma from "@/prisma/client";
 import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
-export async function PATCH(
-  request: NextRequest,
-  { params }: { params: { id: string } }
-) {
+export async function PATCH(request: NextRequest) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
+
+  const url = new URL(request.url);
+  const id = url.pathname.split("/").pop();
+  if (!id)
+    return NextResponse.json({ error: "Missing issue ID" }, { status: 400 });
 
   const body = await request.json();
   const validation = patchIssueSchema.safeParse(body);
@@ -18,8 +20,11 @@ export async function PATCH(
     return NextResponse.json(validation.error.format(), { status: 400 });
 
   const issue = await prisma.issue.findUnique({
-    where: { id: parseInt(params.id) },
+    where: { id: parseInt(id) },
   });
+
+  if (!issue)
+    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
 
   const { assignedToUserId, title, description } = body;
 
@@ -31,9 +36,6 @@ export async function PATCH(
     if (!user)
       return NextResponse.json({ error: "Invalid user." }, { status: 400 });
   }
-
-  if (!issue)
-    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
 
   const updatedIssue = await prisma.issue.update({
     where: { id: issue.id },
@@ -47,20 +49,26 @@ export async function PATCH(
   return NextResponse.json(updatedIssue);
 }
 
-export async function DELETE(
-  request: NextRequest,
-  { params }: { params: { id: string } }
-) {
+
+export async function DELETE(request: NextRequest) {
   const session = await getServerSession(authOptions);
   if (!session) return NextResponse.json({}, { status: 401 });
+
+  const id = request.nextUrl.searchParams.get("id");
+  if (!id)
+    return NextResponse.json({ error: "Missing issue ID" }, { status: 400 });
+
   const issue = await prisma.issue.findUnique({
-    where: { id: parseInt(params.id) },
+    where: { id: parseInt(id) },
   });
 
   if (!issue)
     return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
+
   await prisma.issue.delete({
     where: { id: issue.id },
   });
+
   return NextResponse.json({});
 }
+
